+++ src/emulator/controller.ts
@@ -19,6 +19,7 @@ import {
 } from "./types";
 import { Constants, FIND_AVAILBLE_PORT_BY_DEFAULT } from "./constants";
 import { FunctionsEmulator } from "./functionsEmulator";
+import { parseRuntimeVersion } from "./functionsEmulatorUtils";
 import { DatabaseEmulator, DatabaseEmulatorArgs } from "./databaseEmulator";
 import { FirestoreEmulator, FirestoreEmulatorArgs } from "./firestoreEmulator";
 import { HostingEmulator } from "./hostingEmulator";
@@ -372,7 +373,9 @@ export async function startAll(options: any, noUi: boolean = false): Promise<voi
         ...options.extensionEnv,
       },
       predefinedTriggers: options.extensionTriggers,
-      nodeMajorVersion: options.extensionNodeVersion || options.config.get("functions.runtime"),
+      nodeMajorVersion: parseRuntimeVersion(
+        options.extensionNodeVersion || options.config.get("functions.runtime")
+      ),
     });
     await startEmulator(functionsEmulator);
   }