+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fixes issue where `Authorization` header was missing from callable functions in the emulator (#2459).
+- Improve support for the Node.js 12 (Beta) runtime in the Functions emulator.